rename crossed-eye icon
