+++ projects/packages/videopress/src/client/components/icons/crossed-eye-icon.js
@@ -1,6 +1,6 @@
 import { SVG, Path } from '@wordpress/components';
 
-const privacy = (
+const crossedEye = (
 	<SVG width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
 		<Path d="M8.45414 21.2068L16.4541 3.70678L15.5446 3.29102L13.6676 7.39695C13.1429 7.3024 12.587 7.24988 11.9996 7.24988C9.00203 7.24988 6.82613 8.61747 5.41986 9.95678C4.71735 10.6259 4.19984 11.2937 3.85693 11.7955C3.6851 12.047 3.55604 12.2583 3.46859 12.4095C3.42483 12.4851 3.39141 12.5458 3.36817 12.5891C3.35655 12.6107 3.34747 12.6281 3.34091 12.6407L3.33296 12.6562L3.3304 12.6613L3.3291 12.6639C3.3291 12.6639 3.32878 12.6645 3.9996 12.9999C4.67042 13.3353 4.67016 13.3358 4.67016 13.3358L4.67297 13.3304C4.67621 13.3241 4.68183 13.3134 4.68987 13.2984C4.70594 13.2685 4.73159 13.2217 4.76694 13.1607C4.83769 13.0384 4.94691 12.8591 5.0954 12.6418C5.39312 12.2061 5.84435 11.6239 6.45434 11.043C7.67307 9.88229 9.49718 8.74988 11.9996 8.74988C12.353 8.74988 12.693 8.77247 13.0196 8.81455L12.8834 9.11243C12.6011 9.03897 12.3049 8.99988 11.9996 8.99988C10.0666 8.99988 8.49962 10.5669 8.49962 12.4999C8.49962 13.6966 9.1002 14.753 10.0163 15.3842L7.54467 20.791L8.45414 21.2068Z" />
 		<Path d="M15.7168 9.69537C16.4332 10.0877 17.041 10.563 17.5451 11.0431C18.1551 11.624 18.6064 12.2062 18.9041 12.6419C19.0526 12.8592 19.1618 13.0385 19.2325 13.1608C19.2679 13.2218 19.2935 13.2686 19.3096 13.2985L19.3211 13.3202L19.3265 13.3305L19.3293 13.3359C19.3293 13.3359 19.329 13.3354 19.9999 13C20.6707 12.6646 20.67 12.6632 20.67 12.6632L20.6691 12.6614L20.6665 12.6563L20.6586 12.6408C20.652 12.6282 20.6429 12.6108 20.6313 12.5892C20.6081 12.5459 20.5746 12.4852 20.5309 12.4096C20.4434 12.2584 20.3144 12.0471 20.1425 11.7956C19.7996 11.2938 19.2821 10.626 18.5796 9.95691C17.9765 9.38256 17.2319 8.803 16.3417 8.32834L15.7168 9.69537Z" />
@@ -8,4 +8,4 @@ const privacy = (
 	</SVG>
 );
 
-export default privacy;
+export default crossedEye;
